fix: extracted scenarioProjectList to standalone component. fix: made login form the entry point@
