+++ frontend/src/pages/project-list/index.tsx
@@ -1,0 +1,14 @@
+// pages/index.tsx
+import React from 'react';
+import ProjectManager from '@/components/sub/ScenarioProjectList';
+import RootLayout from '@/components/main/RootLayout';
+
+const ProjectList: React.FC = () => {
+  return (
+    <RootLayout>
+      <ProjectManager />
+    </RootLayout>
+  );
+};
+
+export default ProjectList;
